+++ src/frontend/app/common.ts
@@ -1,1 +1,2 @@
-export type Uuid = string
+export type Uuid = string;
+export type DateTime = number;