+++ src/frontend/app/components/navbar.tsx
@@ -26,7 +26,7 @@ function UserLoginButton() {
         </div>
         <ul 
           tabIndex={0}
-          className="menu menu-sm dropdown-content bg-base-100 rounded-box z-[1] mt-3 w-52 p-2 shadow">
+          className="menu menu-sm dropdown-content bg-base-100 rounded-box z-[1] mt-3 w-52 p-2 shadow border-base-content/10 border-1">
           <li><a>Change</a></li>
         </ul>  
       </div>